Fix admin route params
